Stori/requirements.txt
alembic==1.13.1
exceptiongroup==1.2.0
greenlet==3.0.3
iniconfig==2.0.0
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
